+++ interface.c
@@ -156,7 +156,7 @@ check_iface(struct Interface *iface)
 
 	if ((iface->AdvLinkMTU != 0) &&
 	   ((iface->AdvLinkMTU < MIN_AdvLinkMTU) ||
-	   (iface->if_maxmtu != -1 && (iface->AdvLinkMTU > iface->if_maxmtu))))
+	   (iface->if_maxmtu != -1 && (iface->AdvLinkMTU > (uint32_t)iface->if_maxmtu))))
 	{
 		flog(LOG_ERR,  "AdvLinkMTU for %s (%u) must be zero or between %u and %u",
 		iface->Name, iface->AdvLinkMTU, MIN_AdvLinkMTU, iface->if_maxmtu);
